requirements.txt
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.4.0
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.5
branca==0.4.1
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.4.1
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
Fiona==1.8.13
flake8==3.7.9
folium==0.11.0
future==0.18.2
GDAL==3.0.4
geographiclib==1.50
geopandas==0.7.0
geopy==1.22.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
inflection==0.3.1
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab-widgets==1.0.0
keyring==21.2.1
lazy-object-proxy==1.4.3
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-service==2.3.0
more-itertools==8.3.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
numpydoc==1.0.0
openpyxl==3.0.3
packaging==20.4
pandas==1.0.4
pandasql==0.7.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pg8000==1.19.5
pickleshare==0.7.5
plotly==4.14.3
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2==2.8.6
py==1.8.1
pyarrow==0.17.1
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.5.2
PyNaCl==1.3.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.6.1.post1
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
retrying==1.3.3
rope==0.17.0
Rtree==0.9.4
scipy==1.4.1
scramp==1.4.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
spyder-unittest==0.4.1
SQLAlchemy==1.3.17
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.9
watchdog==0.10.2
wcwidth==0.2.2
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yapf==0.29.0
zipp==3.1.0
